fix: change type x_ratio, y_ratio

+++ src/main/java/com/tikitaka/backend/question/dto/QuestionCreateRequest.java
@@ -19,8 +19,8 @@ public class QuestionCreateRequest {
     private Boolean isAnonymous;
 
     @JsonProperty("x_ratio")
-    private Double xRatio;
+    private Float xRatio;
 
     @JsonProperty("y_ratio")
-    private Double yRatio;
+    private Float yRatio;
 }
